Database suffix generator
The default suffix generator uses DateTime.Now.Ticks, this is normally
ok, until you run in parallel with a unit test runner that decides
to put all the localdb files in the same directory. This can cause an
overlap that leads to test failures.

The suffix generator allows you to inject a unique suffix based on any
generator you can think of: Guid, DateTime, etc.

#16